Add cancel module
Add move AsyncFd::cancel_previous, AsyncFd::cancel_all, the futures
types Cancel and CancelOp, and CancelResult into it.

Will also be the future home of the Cancel trait.
