+++ tests/async_fd/net.rs
@@ -9,7 +9,8 @@ use std::os::fd::{AsFd, AsRawFd, BorrowedFd};
 use std::pin::Pin;
 use std::ptr;
 
-use a10::io::{CancelResult, ReadBufPool};
+use a10::cancel::CancelResult;
+use a10::io::ReadBufPool;
 use a10::net::{Accept, MultishotAccept, MultishotRecv, Recv, Send, Socket};
 use a10::{Extract, Ring};
